+++ lib/td_se_web/controllers/fallback_controller.ex
@@ -1,0 +1,38 @@
+defmodule TdSeWeb.FallbackController do
+  @moduledoc """
+  Translates controller action results into valid `Plug.Conn` responses.
+
+  See `Phoenix.Controller.action_fallback/1` for more details.
+  """
+  use TdSeWeb, :controller
+
+  # This clause handles errors returned by Ecto's insert/update/delete.
+  def call(conn, {:error, %Ecto.Changeset{} = changeset}) do
+    conn
+    |> put_status(:unprocessable_entity)
+    |> put_view(json: TdSeWeb.ChangesetJSON)
+    |> render(:error, changeset: changeset)
+  end
+
+  def call(conn, {:error, :unprocessable_entity, message}) do
+    conn
+    |> put_status(:unprocessable_entity)
+    |> put_view(json: TdSeWeb.ErrorJSON)
+    |> render(:error, message: message)
+  end
+
+  def call(conn, {:error, :forbidden}) do
+    conn
+    |> put_status(:forbidden)
+    |> put_view(json: TdSeWeb.ErrorJSON)
+    |> render(:"403")
+  end
+
+  # This clause is an example of how to handle resources that cannot be found.
+  def call(conn, {:error, :not_found}) do
+    conn
+    |> put_status(:not_found)
+    |> put_view(html: TdSeWeb.ErrorHTML, json: TdSeWeb.ErrorJSON)
+    |> render(:"404")
+  end
+end