+++ lib/td_se_web/controllers/search_controller.ex
@@ -2,29 +2,12 @@ defmodule TdSeWeb.SearchController do
   @moduledoc """
   Controller module for global search engine.
   """
-  use PhoenixSwagger
   use TdSeWeb, :controller
   alias TdSe.GlobalSearch
-  alias TdSeWeb.SearchResultsView
-  alias TdSeWeb.SwaggerDefinitions
 
   @aliases Application.compile_env(:td_se, :index_aliases)
 
-  def swagger_definitions do
-    SwaggerDefinitions.global_search_definitions()
-  end
-
-  swagger_path :global_search do
-    post("/global_search")
-    description("Search for all of our indexes under the whole search space")
-    produces("application/json")
-
-    parameters do
-      search(:body, Schema.ref(:GlobalSearchRequest), "Search query and filter parameters")
-    end
-
-    response(200, "OK", Schema.ref(:GlobalSearchResponse))
-  end
+  action_fallback TdSeWeb.FallbackController
 
   def global_search(conn, params) do
     claims = conn.assigns[:current_resource]
@@ -70,10 +53,6 @@ defmodule TdSeWeb.SearchController do
 
     conn
     |> put_resp_header("x-total-count", "#{total}")
-    |> put_view(SearchResultsView)
-    |> render(
-      "global_search_results.json",
-      global_search_results: global_search_results
-    )
+    |> render(:index, search_results: global_search_results)
   end
 end